first iteration of a freely moving vertex model
This model is not strictly speaking a vertex model.
Attraction only occurrs between vertices of the cells and not the cell
boundaries (as it should be realistically).
In addition, the algorithms provided are currently only a first working
prototype. We use the ray-casting algorithm for point-in-polygon
detection, which is itself an efficient algorithm but it can surely be
implemented more efficiently.